chore: bump alloy 0.4 (#127)
<!--
Thank you for your Pull Request. Please provide a description above and
review
the requirements below.

Bug fixes and new features should include tests.

Contributors guide:
https://github.com/alloy-rs/core/blob/main/CONTRIBUTING.md

The contributors guide includes instructions for running rustfmt and
building the
documentation.
-->

<!-- ** Please select "Allow edits from maintainers" in the PR Options
** -->

## Motivation

<!--
Explain the context and why you're making that change. What is the
problem
you're trying to solve? In some cases there is not a problem and this
can be
thought of as being the motivation for your change.
-->

## Solution

<!--
Summarize the solution and provide any necessary context needed to
understand
the code change.
-->

## PR Checklist

- [ ] Added Tests
- [ ] Added Documentation
- [ ] Breaking changes

+++ crates/alloy/consensus/src/transaction/envelope.rs
@@ -387,7 +387,7 @@ mod tests {
     #[cfg(feature = "serde")]
     fn test_serde_roundtrip_deposit() {
         let tx = TxDeposit {
-            gas_limit: u128::MAX,
+            gas_limit: u64::MAX,
             to: TxKind::Call(Address::random()),
             value: U256::MAX,
             input: Bytes::new(),
